Remove annotate
The annotations are too easy to out of sync with the actual DB structure.
This means it is likely a developer reading them from the file head could
get wrong information about the modal fields. We should read db/schema.rb
instead as that file is guaranteed to always updated.